requirements.txt
mcp>=1.0.0
# asyncio
# dataclasses
# pathlib
# re
# datetime
# json
typing
openai>=1.0.0
openai-agents
anthropic>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
weave
